fix tests & migrate them to ts
