+++ project/server/package.json
@@ -27,9 +27,12 @@
 	"devDependencies": {
 		"@types/bcryptjs": "^2.4.2",
 		"@types/express": "^4.17.13",
+		"@types/jest": "^28.1.4",
 		"@types/jsonwebtoken": "^8.5.8",
 		"@types/node": "^18.0.0",
+		"@types/supertest": "^2.0.12",
 		"jest": "^28.1.2",
+		"ts-jest": "^28.0.5",
 		"tslint": "^6.1.3",
 		"typescript": "^4.7.4"
 	}
