+++ project/server/src/config/db.ts
@@ -1,7 +1,11 @@
-import mongoose from "mongoose";
+import mongoose from 'mongoose';
 
 const connectDB = async () => {
 	const conn = await mongoose.connect(process.env.MONGO_URI);
 };
 
-export default connectDB;
+const disconnectDB = async () => {
+	await mongoose.disconnect();
+};
+
+export { connectDB, disconnectDB };
