+++ project/server/src/routes/gistRoutes.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import { protect } from '../middleware/authMiddleware.js';
+import { protect } from '../middleware/authMiddleware';
 
 import {
 	getPublicGists,
